sheep: avoid atomic operations against uint64_t/int64_t variables
On 32 bit architectures, atomic operations against 64 bit integers are
not allowed.  This patch uses more proper types for them.

+++ sheep/object_cache.c
@@ -48,7 +48,7 @@
 
 
 struct global_cache {
-	uint64_t cache_size;
+	uint32_t cache_size;
 	int in_reclaim;
 	struct cds_list_head cache_lru_list;
 };
@@ -515,8 +515,8 @@ static int do_reclaim_object(struct object_cache_entry *entry)
 		goto out;
 	}
 
-	dprintf("oid %"PRIx64" reclaimed successfully, cache_size: %ld\n", oid,
-		uatomic_read(&sys_cache.cache_size));
+	dprintf("oid %"PRIx64" reclaimed successfully, cache_size: %"PRId32"\n",
+		oid, uatomic_read(&sys_cache.cache_size));
 	del_from_object_tree_and_list(entry, &oc->object_tree);
 out:
 	pthread_rwlock_unlock(&oc->lock);
